+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Added:
 Fixed:
 * Issue #6: platform doors and panels breaking fix
 
+Changed:
+* The style changer now has damage values
+
 0.7
 ---